[MOD] BOJ-15553-난로.cpp

+++ BOJ/그리디 알고리즘/15553-난로.cpp
@@ -21,7 +21,7 @@ int main(void) {
 
 	sort(subs, subs + N - 1);
 
-	long long ans = K;
+	int ans = K;
 
 	for (int i = 0; i < N - K; ++i)
 		ans += subs[i];
